Refactor out RequestPopup and ClusterMapLayers
