+++ src/webapp/components/ClusterMap.js
@@ -1,9 +1,6 @@
-import React, { useState } from "react";
-import { Layer, Source, MapContext, Popup } from "react-mapbox-gl";
+import React from "react";
+import { Source } from "react-mapbox-gl";
 import { LngLat } from "mapbox-gl";
-import Typography from "@material-ui/core/Typography";
-import Box from "@material-ui/core/Box";
-import Link from "@material-ui/core/Link";
 import { findBounds } from "../helpers/mapbox-coordinates";
 import {
   CROWN_HEIGHTS_BOUNDS,
@@ -11,6 +8,7 @@ import {
 } from "../helpers/map-constants";
 import BasicMap from "./BasicMap";
 import { QuadrantsLayers } from "./QuadrantMap";
+import ClusterMapLayers from "./ClusterMapLayers";
 
 const makeBounds = geoJsonData => {
   const lnglats = geoJsonData.features.map(f => {
@@ -27,8 +25,6 @@ const makeBounds = geoJsonData => {
 };
 
 const ClusterMap = ({ geoJsonData, containerStyle = {} }) => {
-  const [popup, setPopup] = useState();
-
   if (!geoJsonData) {
     return null;
   }
@@ -50,116 +46,7 @@ const ClusterMap = ({ geoJsonData, containerStyle = {} }) => {
           clusterRadius: 30
         }}
       />
-      <MapContext.Consumer>
-        {map => (
-          <>
-            <Layer
-              sourceId="clusterSource"
-              type="circle"
-              id="clusters"
-              filter={["has", "point_count"]}
-              paint={{
-                "circle-color": [
-                  "step",
-                  ["get", "point_count"],
-                  "orangered",
-                  5,
-                  "#d03800",
-                  10,
-                  "#a22b00"
-                ],
-                "circle-radius": [
-                  "step",
-                  ["get", "point_count"],
-                  20,
-                  5,
-                  30,
-                  20,
-                  40
-                ]
-              }}
-              onClick={e => {
-                const features = map.queryRenderedFeatures(e.point, {
-                  layers: ["clusters"]
-                });
-                const clusterId = features[0].properties.cluster_id;
-                map
-                  .getSource("clusterSource")
-                  .getClusterExpansionZoom(clusterId, (err, zoom) => {
-                    if (err) return;
-
-                    map.easeTo({
-                      center: features[0].geometry.coordinates,
-                      zoom
-                    });
-                  });
-              }}
-            />
-
-            <Layer
-              id="cluster-count"
-              type="symbol"
-              sourceId="clusterSource"
-              filter={["has", "point_count"]}
-              layout={{
-                "text-field": "{point_count_abbreviated}",
-                "text-font": ["Arial Unicode MS Bold"],
-                "text-size": 14
-              }}
-              paint={{
-                "text-color": "#ffffff"
-              }}
-            />
-
-            <Layer
-              id="unclustered-point"
-              type="circle"
-              sourceId="clusterSource"
-              filter={["!", ["has", "point_count"]]}
-              paint={{
-                "circle-color": "orangered",
-                "circle-radius": 6,
-                "circle-stroke-width": 1,
-                "circle-stroke-color": "#e73e00"
-              }}
-              onClick={e => {
-                const meta = JSON.parse(e.features[0].properties.meta);
-                setPopup({
-                  lngLat: e.lngLat,
-                  meta
-                });
-              }}
-            />
-          </>
-        )}
-      </MapContext.Consumer>
-
-      {popup && (
-        <Popup
-          coordinates={popup.lngLat}
-            offset={{
-            "bottom-left": [12, -38],
-            bottom: [0, -38],
-            "bottom-right": [-12, -38]
-          }}
-        >
-          <Box>
-            <Typography variant="h6">{popup.meta["First Name"]}</Typography>
-            <Typography variant="body1">
-              {popup.meta["Cross Street #1"]}
-              {' and '}
-              {popup.meta["Cross Street #2"]}
-            </Typography>
-            <Link
-              href={`https://crownheightsmutualaid.slack.com/archives/${popup.meta.slackChannelId}/p${popup.meta.slackTimestamp}`}
-              target="_blank"
-            >
-              See details on Slack
-            </Link>
-            <Typography variant="body2">Request code: {popup.meta.Code}</Typography>
-          </Box>
-        </Popup>
-      )}
+      <ClusterMapLayers />
     </BasicMap>
   );
 };
